move split function to utilities, improve documentation
